Added YouTubeSearchAndWatchHistory link GH-60
Started setting up the architecture to allow linking data from multiple sources into a single visualisation. The first example that this commit contributes is linking youtube search history with youtube watch history to see which searches led to which videos being watched. This is not complete yet, and hasn't been added to the app yet. Mostly just additions to lib.

Signed-off-by: Brian Evans <ebrian101@gmail.com>